requirements.txt
asgiref
certifi
dj-database-url
Django
django-ckeditor
django-cloudinary-storage
django-compressor
django-cors-headers
django-environ
django-debug-toolbar
django-js-asset
django-recaptcha==3.0.0
djangorestframework
gunicorn
g4f==0.0.3.4
idna
psycopg2
Pillow
pytz
requests
redis
six
sqlparse
typing_extensions
urllib3==1.26.12
whitenoise
